requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
diskcache==5.2.1
et-xmlfile==1.1.0
idna==2.10
lxml==4.6.3
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
xlrd==2.0.1
